back_flask/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
bcrypt==4.1.2 